refactor(graphql-client): split lib.rs into modular API files
