+++ crates/iota-sdk-graphql-client/src/api/mod.rs
@@ -1,0 +1,19 @@
+// Copyright (c) Mysten Labs, Inc.
+// Modifications Copyright (c) 2026 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! API implementations for the GraphQL client.
+
+mod balance;
+mod checkpoints;
+mod coins;
+mod dry_run;
+mod dynamic_fields;
+mod epochs;
+mod events;
+mod iota_names;
+mod network;
+mod objects;
+mod package;
+mod packages;
+pub(crate) mod transactions;
